feat : Clock util function 구현
