+++ src/Utils/Clock.js
@@ -1,0 +1,19 @@
+export function Clock(setClock) {
+  const today = new Date()
+  const utc = today.getTime() + today.getTimezoneOffset() * 60 * 1000
+  const timeGapKr = 9 * 60 * 60 * 1000
+  const koreaTime = new Date(utc + timeGapKr)
+  const koreaTimeFormat = new Intl.DateTimeFormat('ko-KR', {
+    dateStyle: 'full',
+    timeStyle: 'short',
+  }).format(koreaTime)
+  const usTimeFormat = new Intl.DateTimeFormat('en-US', {
+    dateStyle: 'full',
+    timeStyle: 'short',
+  }).format(today)
+  setClock((prev) => ({
+    ...prev,
+    koreaClock: koreaTimeFormat,
+    usClock: usTimeFormat,
+  }))
+}
